src/store/userSlice.ts
import { UserLoginData } from "@models/userModels";
import { PayloadAction, createSlice } from "@reduxjs/toolkit";

interface UserState {
  user: UserLoginData | null;
}

const initialState: UserState = {
  user: null,
};

const userSlice = createSlice({
  name: "user",
  initialState,
  reducers: {
    setUser: (state, action: PayloadAction<UserLoginData>) => {
      state.user = action.payload;
    },
    clearUser: (state) => {
      state.user = null;
    },
  },
});

export const { setUser, clearUser } = userSlice.actions;
export const userReducer = userSlice.reducer;
